added user login thunk
